Some BugFix V0.10.1 (wrong voltage for HoTT, current casting)
Former-commit-id: e1a06391745cd96df2b9c500975057649cd77fe0 [formerly bb7ef0f7afd1431449f00e2e114112a7d2c341c5]
Former-commit-id: 8d622c23f64088f67998342af730d2851c70b0c3

+++ Firmware/src/module_SPort.c
@@ -437,7 +437,7 @@ static void RX_Callback(uint8_t* id)
 				switch (session->rxPacket.Id)
 				{
 				case SENSOR_CURR:
-					simpleTelemtryData.current = SPORT_DATA_U32(session->rxPacket) / 10;
+					simpleTelemtryData.current = ((float)SPORT_DATA_U32(session->rxPacket)) / 10;
 					break;
 				case SENSOR_CELLS:
 				{
